Auto-completion: Faster fits parse, use 'local'
With this commit, I added the '-q' option to 'astfits' program so it will
print less information, hence, it looks might be a bit faster.

Since we can use the 'local' declaration, I removed the 'unset' at the
end of functions and declared their variables as 'local' instead.

Also, there was a misleading typo in comments about return value of the
_gnuastro_autocomplete_plaintext_is_table function. It said the function
would return '1' if the given file is a table, on the contrary, it
returns '0', as it should by the conventions.