Auto-completion: Debug -i, improve performance
Until now, autocompletion checked if there is a valid FITS file or a
plaintext table available while also the '-i' or '--information' option
is called. In this case, it would print a warning message that it is safe
to press ENTER since all other options will be disabled by the 'table'
program. However, this behavior broke in previous commits. This commit
fixes the problem.

Also, previously the 'awk' command in the '..._last_table' function would
print all the words in the command line -- named as 'token'. This
includes the short and long options. The problem is that bash is later
instructed to check if each 'token' is a valid file -- [ -f $token]. As
Mohammad pointed out before, this is a computationaly exhaustive process.
So, we would only need to check tokens that we already know are not the
program options! These tokens would start with a dash '-'. Hence, I have
added a regex condition check in the 'awk' program to take care of that.
Now, it will only print tokens NOT starting with a '-' sign.