Auto-completion: Call ..._last_table where needed
Until now, the '..._last_table' function ran everytime. This function is
responsible for populating the $last_table variable. Which means there
has been a valid FITS or plaintext table file called in the commandline.
Since we do not need to check for this everytime the script runs, I have
moved it to more specific places where we need to check the $last_table
variable. This might help reduce overhead.

Also, I noticed and removed the extra space between the command prompt
and the user commands.