Auto-completion: Refactor ..._is_table function
With this commit, I replaced 'return 0' (or 1) with 'break' as it used to
be. Because we are not using the function as a boolean query. However, we
can se if there is a valid FITS or plaintext file withing the command
line by checking if the $last_table is populated: [ x$last_table != x ]

Also, I replaced an indented 'if' with 'elif' for better code readablity.