Query: added NED's Galactic extinction calculator
Until now, only one of NED's very rich dataset was available within
Query. This was because only that one table followed the general Table
Access Protocol (TAP) of the Virtual Observatory standard.

With this commit, NED's galactic extinction calculator for a certain
coordinate is now also available. For this feature, we use NED's custom CGI
interface, not the standard TAP. The format of the CGI URL was kindly
provided by Joseph Mazzarella from NED.

The only issue is that NED provides a VOTable which Gnuastro doesn't yet
support. So the necessary shell command to convert the VOTable data into a
FITS table is also given in the manual.

This commit was done with the help of Joseph Mazzarella from NED.

This completes task #15911.