Book: example of a good commit message added
Until now the "Commit guidelines" section of the "Developing" chapter had
just listed the suggestions and tips. But without a useful example it was
hard for new users to adopt.

With this commit, an example text has been added that actually describes
the logic within each paragraph of the commit message to help clarify the
ideal commit message to new developers.