MakeCatalog: reported surface brightness limit accounting for area
Until now the surface brightness limit reported in MakeCatalog's output
headers didn't account for the area that it was calculated over (value of
the '--sfmagarea' option). The equation of the surface brightness limit
within the book also had a similar problem.

With this commit, both these issues have been corrected in the source of
MakeCatalog and in the book. In particular a new "Image surface brightness
limit" section has been added to the third tutorial to discuss the concept
with real examples on the data. In process many other necessary corrections
were made as listed below.

 - Book: the "Brightness, flux and magnitude" sub-section was brought to
   the MakeCatalog section. Until now, it was under the MakeProfiles
   section.

 - MakeCatalog now has an extra '--upperlimitsb' measurement which will
   return the upper-limit surface brightness level for the labeled object.

 - MakeCatalog writes the calculated surface brighntess values as FITS
   keywords into the output. Until now, it was in the form of
   'COMMENTS'. In order to do this, some major re-structuring of the
   relevant functions was necessary.

 - In the library, I noticed that the 'txt_write_keys' function didn't have
   a working scenario for full comment lines so it was added.