Library (units.h and arithmetic.h): new counts-to-mag function
Until now, when we wanted to convert counts to magnitudes anywhere in
Gnuastro we would use the raw equation. This could be buggy (mistakenly
missing a negative, or etc!). On the other hand, for a Gnuastro user,
converting counts to magnitudes for a certain count level required a
relatively long and inconvenient AWK or Arithmetic command!

With this commit, the 'gal_units_counts_to_mag' function has been added
inside of Gnuastro's library. Thanks to that function, a new operator has
been added to the Arithmetic library (and thus the Arithmetic program or
Column Arithmetic in Table). Also, with the function in place, all the
places where we had the raw magnitude estimations (using '-2.5*log10(B)+Z')
are replaced with a call to this function and thus the code has been
cleaned up.