Book: Gnuastro manifesto citing the Maneage paper
Until now the second section of Gnuastro's book was simply called "Science
and its tools", but this wasn't too clear on its purpose. Also, within the
text, previously we had only given the Gitlab page for the very old version
of Maneage. But since then a paper has been submitted on Maneage which is
much more clear and purposefully written (arXiv:2006.03018).

With this commit, the section is now called "Gnuastro manifesto: science
and its tools" to be more explicit about the purpose of this chapter. From
the Merriam-Webster dictionary, a manifesto is "a written statement
declaring publicly the intentions, motives, or views of its issuer".

Also, I noticed that the famous quote from Buckheit & Donoho (1996) fit
very nicely in the part where we discussed how English papers aren't enough
to convey the science result, so it has been added.

Finally the double quotations have been removed from the word 'puzzle' when
referring to Kuhn's usage of the term to make it more readable. It is clear
from the footnote immediately after it that this comes from Kuhn's book.