astscript-sort-by-night: make check tests done on certain files
Until now, the tests of the 'astscript-sort-by-night' script would run on
all the FITS files in the tests directory! As a result, when a problematic
file from a previously failed test existed, it would crash!

With this commit, instead of running blindly on all '*.fits' files, the
test will only run on files that have been successfully made by the
prerequisites of this test.