All programs: --wcslinearmatrix to select PCi_j and CDi_j in output
Until now, Gnuastro's programs would only write their output WCS linear
matrix keywords in the 'PCi_j' formalism (which was the default of
WCSLIB). For the TPV distortion, it would use the 'CDi_j' formalism
(because of a problem in WCSLIB for using 'PCi_j' with the TPV, TNX and ZPX
distortions) until it is fully implemented in WCSLIB and print a warning
about it. However, in a communation with Mark Calabretta (author of
WCSLIB), he informed me that implementing the correction would be very
complicated (in the P.S).

Since the TPV distortion will not work with the 'PCi_j' within WCSLIB any
more, with this commit, the warning has been removed. Furthermore, with
this commit, since the low-level infrastructure already existed to write
the WCS as a 'CDi_j', I generalized it as a common option to all programs:
--wcslinearmatrix. The default mode is still the 'PCi_j', but in some cases
it may be necessary for the users to write the output matrix as 'CDi_j'.

Since this is a pretty low-level operation, the best to implement it was to
use this in the part where we read the WCS structure and internally tweak
the 'wcsprm' so the 'PCi_j' and 'CDi_j' matrices are identical ('CDELTi'
would be 1.0). Then when writing the WCS structure and convert the 'PCi_j'
keywords into 'CDi_j' and delete the 'CDELTi' keywords. Therefore it was
necessary to give the 'wcslinearmatrix' option to the 'gal_wcs_read',
'gal_wcs_read_fitsptr'. Also, to 'gal_wcs_to_cd' function is now a public
function in the library (to let programs also use this option directly: in
particular MakeProfiles).

P.S. Note from WCSLIB's author: "I had also said in a previous email in
that thread that I was working on a fix.  In fact, having spent quite a bit
of time on it, I realised that I was creating a monster (without a
compelling reason to do so), and ended up backing out of it completely.
Complications arose in handling TPV in the case where there were additional
axes.  The bookkeeping required to keep track of the special handling of
PCi_j + CDELTi for TPV (and TNX and ZPX) bled over into other parts of
WCSLIB that didn't ought to have to deal with such issues, and it made the
code look and feel ugly.  (I have often had to complicate the WCSLIB code
to handle past non-standard usage, and it just seems to defeat all of the
effort that went into developing the WCS standard!)  I ended up adding a
warning against splitting CDi_j into PCi_j + CDELTi for TPV, (and TNX, and
ZPX)."