Fits: --keyvalue option only doesn't print name with --quiet
Until now, '--keyvalue' wouldn't include the file name as a first column
with only one input. When there was more than one input, it would always
print the file names as a first column. As a result, in a general script
(that needs consistancy and you don't know the number of inputs a-priori),
if the input was only a single file it would require extra work by the
script writer (which is not good!).

With this commit, '--keyvalue' will also print the file name when there is
only one file to be consistent. To avoid printing the filename (when there
is only a single file: the previous behavior), you should use '--quiet'.