Fits: new --wcscoordsys option to convert coordinate systems
Until now there was no easy way to convert the coordinates of an image's
WCS. I had previously asked about this from Mark Calabretta (author of
WCSLIB) about this and he kindly added the 'wcsccs' function for this job
to WCSLIB 7.5 (released on March 21st, 2021).

With this commit, the 'gal_wcs_coordsys_convert' wrapper function (over
'wcsccs') has been added to Gnuastro's library for this job (besides a
function to read the user's desired coordinate system and another function
to set the pole coordinates for the various conversions). The pole
coordinates were found (inspired from the examples in WCSLIB's manual) from
the NED coordinate calculator.

With the library functions in place, it was very easy to add the necessary
steps in the Fits program for the new '--wcscoordsys' option. It will parse
the input WCS structure, find its coordinate system, and convert it to the
user's desired system.