ConvertType: new --globalhdu to only specify HDU once
Until now, when there was more than one FITS input, it was necessary to
call the '--hdu' option multiple times (once for every FITS file). But this
was very annoying when all the three HDUs were the same.

With this commit, following the model of Arithmetic and '--globalhdu' (or
'-g') option has been added to ConvertType and behaves similarly: when
given, the same HDU will be read from all the images.