ConvertType: warning added when images are not aligned
Until now, ConvertType would only check the pixel size of the input
images. However, it could happen that the pixel sizes are the same, but the
WCS alignment of the images is different.

With this commit, a check has been added at the start of ConvertType for
this: if the input images have WCS, their central coordinate is checked in
WCS coordinates and if it different by more than one pixel, ConvertType
will print an informative warning about it. To provide a fast solution, a
section has also been added to the ConvertType section of the Gnuastro book
with a small script that people can just copy-paste to fix the problem.

This issue was found with the help of Raúl Infante Sainz.