MakeProfiles: NaN value in custom profile will be NaN in output
Until now, when MakeProfiles was given a custom profile with a NaN value it
would put a 0 for those pixels. This didn't conform with the documentation
(which claimed that any value given to the third column will be
used). Also, it could cause a confusion for radii that are in the custom
profile (but with a value of NaN), and radii that are not in the custom
profile's range.

This wrong feature was manually inserted at the end of the custom profile
function: we would check the output value, and if it was NaN, we would
insert a 0 instead!

With this commit, the feature the raw value from the user's catalog is used
and the 'out' value for the custom profile function is initialized to
'0.0', so the function has also been polished/cleaned for this.

This has also been reflected inside the documentation. Also, I noticed that
in the book, two commands were given to construct the custom table, one
with Gnuastro's Table and one with AWK. But because they were under each
other in a single code block, it could confuse the readers on a first
reading. Also, the extra explanation made the description long and
confusing. So the AWK command has been removed with this commit.

This issue was reported by Carlos Morales Socorro.