MakeProfiles: better option explanation with --help and -P
Until now, when run with '--help' or '-P', some of the option descriptions
were just slightly long enough to force a breaking of the line into
two. This made it hard to read the option values.

With this commit, these options have been just slightly shortened to be
easier to read. In the process, I also fixed the description of the
'--minmapsize' option to fit its long value within the fixed length.