Library (wcs.h): WCS linear matrix converted at start
Until now, the recent change where we allow the user was given the ability
to select the linear matrix (Commit 20f1d41d) had an un-expected
side-effect: when the input had a CD matrix, the output would also have a
CD matrix even when '--wcscoordsys' was set to 'pc'.

With this commit, the relevent 'else' has been added to correct this missed
point.

This issue was found with the help of Zahra Sharbaf.