Book: minor typo corrected
Until now, the "start of" statement was repeated two times inside a comment
of a code block in the first tutorial. While looking at the PDF to find
this problem, I also noticed that the output of 'mkprof -P' didn't include
the new '--fcol' custom profile that was recently added and also the
description of '--ccol' was going beyond the text width.

With this commit, both issues have been fixed.

This repeating of "start of" was found by Sepideh Eskandarlou.