Crop: added option to crop polygon from ds9 file
Until now, to define a cropping polygon, it was necessary to pass the
polygon vertice coordinates to the '--polygon' option on the
command-line. However, polygons are usually defined based on the larger
image and trying to visually identify their vertices by hovering your mouse
over them and writing down the coordiantes is very annoying and prone to
serious bugs. The need for this had already been discussed in task #14409
(Polygon vertices from a DS9 region file).

With this commit, work is starting on implementing this feature through a
new '--polygonname' option.

NOTE: commit message edited by Mohammad Akhlaghi.