Book: added notice on single-quotes for PDF outputs
Until now, we hadn't explicitly mentioned reminded readers that the printed
single quote in PDF manual is the wrong character to write or to
copy-paste.

With this commit, a cartouche box has been added on top of the first AWK
command in the manual where they occur to remind unsuspecting new
users. Also an explanation has been added in the "Conventions" section of
the first chapter.

This issue was reported by Sepideh Eskandarlou.