Book: Quantifying measurement limits divided into subsections
Until now, the various measurement limit measurements were put as an item
in a table. Since they are growing and their number is also increasing,
this made them hard to find and read in this long section.

With this commit each of the various methods is now given its own
sub-sub-section to help directly referencing only one of them and also let
the reader breathe between them.

Also, while discussing the surface brightness limit with Juan Miro he
mentioned that the analogy with muddy water helped him a lot in
understanding the concept. However, that analogy was recently removed
(because I thought it may not be useful). So with this commit, that
paragraph has been re-inserted into the "Surface brightness limit of image"
part.