Library (wcs.h): initialized variable to avoid compiler warning
Until now, the 'gal_options_read_wcslinearmatrix' function (which is in
charge of reading the value to the new '--wcslinearmatrix' option) would
produce an "may be used uninitialized" warning on some compilers for the
'value' variable. In practice, if there is a problem in reading the user's
value control would never reach that stage and the program will abort, but
the compilers don't know about this higher-level logic.

With this commit the variable 'value' has been initialized to the "invalid"
identifier to avoid this compiler warning and help in having a clean
compilation output.

While doing this, I noticed that the invalid value for the linear
transformation matrix ('GAL_WCS_LINEAR_MATRIX_INVALID') was wrongly named,
so it has been corrected and added into the book. I also noticed that the
comment indendation for the 'GAL_WCS_COORDSYS' set of macros didn't align
with the others, so that has also been corrected for readability.

This issue was reported by Carlos Morales Socorro.