astscript-ds9-region: new name for astscript-make-ds9-reg
Until now, this installed script started with 'make-', therefore making it
hard to remember or find the script (especially on the command-line with
TAB).

With this commit, the name of the script starts with the main point of the
script (which is 'ds9') and it also fully spells 'region' to help in
readability/usability.