bootstrap.conf: no more need to fix in ax_pthread.m4
Until now, the 'ax_pthread.m4' check of Autoconf Archives used 'as_echo'
that would create trouble with the newer versions of GNU Autoconf. In a
recent pull of their commits, I noticed that this issue has been fixed and
there is no more need for the manual correction here.

With this commit, that part of 'bootstrap.conf' has been removed.