NEWS: minor edits in the parts for next stable release
Until now, the various components of the NEWS file were written separately
with each new feature or bug fix. But after an alpha release based on the
previous commit, some improvements could be found.

With this commit, I went over the full text and implemented some edits to
make it more easier to read, highlight somethings and also do a
spell-check.