Table: added the --polygonname option
Until now, the new ds9 library allowed the option --polygonname in the Crop
program to be extended to the Table program.

With this commit, using the new ds9 library, the --polygonname option that
was first implemented in the Crop program is now also in the Table program.

NOTE: this commit message was re-formated by Natáli Anzanello to follow the
convention about the message-line size.