Book: new --polygonname option and new ds9 library added
Until now, the --polygonname option was just in the code of the Table and
Crop programs. Also, the new ds9 library that is essential to this new
option was not documented.

With this commit, the --polygonname option is also documented in the Table
and Crop sections of the book. The new ds9 library is also documented.