Library (arithmetic.h): new operators to add noise
Until now, the only way to add noise was with the MakeNoise program. This
had two limitations: 1) it didn't work on tables, 2) it wasn't easy to
blend with other types of operations on the pixels.

With this commit, the core operation of MakeNoise has now been moved into
the arithmetic library as two operators ('mknoise-sigma' and
'mknoise-poisson'). This enables both the features above.

However, it doesn't yet support instrumental noise, but that can be added
later, so its still too early to completely remove MakeNoise! Although it
isn't too far out of sight now: possibly with a third operator, I just
don't have enough time to implement it now.