Building: initialized some variables to avoid compiler warnings
Until now, some of the variables in Gnuastro had the potential to trigger
"may be used un-initalized" warnings. I just noticed them recently on my
own computer! Looking into the code, these could never cause a problem (the
program would abort with an informative message if they weren't set), but
well, this is too high-level for compiler, so it prints these warnings.

With this commit those variables (all were pointers) are initalized to
'NULL' and the warnings have been silenced.