Library (arithmetic.h): envseed implemented as a flag
Until now the 'envseed' setting was passed to the arithmetic operators that
needed it as an extra dataset. But this was a very dirty and ugly solution
for a very simple on-off situation (it needed the allocation of a dataset,
and the subsequent checking of all the properties).

With this commit, the 'envseed' setting is now defined as a new bit-flag in
the 'flags' argument of the 'gal_arithmetic' function. This makes the code
much more elegant, easier to read/parse and debug. I had completely forgot
about flags when implementing the recently added 'mknoise-*' operators!

In addition to the new 'GAL_ARITHMETIC_FLAG_ENVSEED', I also defined a new
'GAL_ARITHMETIC_FLAG_QUIET' flag (to pass the '--quiet' option from the
programs). Thanks to this, the new 'mknoise-*' operators now print the
random number generator settings by default and this can be disabled with
the '--quiet' option.

In the process, I also done some better management of the arithmetic
library flags:

 - Until now, the arithmetic flag macros had this format:
   'GAL_ARITHMETIC_*'. So it wasn't immediately clear that they are
   flags. With this commit, they now have this format:
   'GAL_ARITHMETIC_FLAG_*'.

 - In a few places, instead of the 'GAL_ARITHMETIC_FLAGS_BASIC', I was
   actually using the three flags that it points to. So to simplify things,
   I am just using the single wrapper macro now.