Library (fits.h): FITS files can be recognized without suffix
Until now, Gnuastro identified FITS files based on their recognized suffix
(for example '.fits', '.fit', or etc). However, in some situations, it may
happen that the file has no suffix, or a recognized suffix. So until now,
users would have to find a hack (for example make a symbolic link to the
file, and set the link name to have a '.fits' suffix). But this was an
annoying step, therefore having this step helps the usability of Gnuastro.

With this commit, a new 'gal_fits_file_recognized' function has been added
to Gnuastro's FITS library and used by all the programs when checking their
inputs. This function will first check the suffix and when the suffix is
not recognized (or doesn't exist!), it will attempt to open the file with
CFITSIO to make sure it is a FITS file. Note that it will just "open" the
file, it won't read anything (beyond checking that it is indeed a FITS
file) so it shouldn't have too much of a footprint.

Also, a more general 'gal_array_file_recognized' function has been added
for programs that can theoretically accept other formats (like JPEG or
TIFF), but this feature is currently only implemented for FITS.

Initially, I also wanted to implement this for similar functions in TIFF
and JPEG, but ended up not doing so because of the peculiarities of those
libraries that I have forgot and need more time to remember. But those
formats are rarely used as input into Gnuastro anyway, and by normal
convention, they usually do have suffixes.

This feature was implemented while working with Clotilde Laigle and Leigh
Smith.