Crop and Table: --polygon now accepts filenames
Until now, we used the '--polygonfile' option to specify the DS9 region
filename containing the polygon.

With this commit, the '--polygon' option also accepts a filename of a DS9
region file. Therefore, the '--polygonfile' option was removed. To do this,
a check in 'gal_options_parse_colon_sep_csv' (in 'lib/options.c') was added.