radial-profile: removing the set-x option used for debuging
Until this commit, the option 'set-x' was used during the development of
the code to debug. It prints the raw code at the same time than the output
of the code while it is being executed. However, for the final version we
do not want to have this. With this commit, this option has been removed.
In addition to this, very minor corrections and descriptions have been
included.