Book: adjusted --polygon option
Until now, there was a '--polygonfile' option used to give a filename from
a DS9 region file. It was removed because the '--polygon' option now
accepts the filename.

With this commit, the '--polygonfile' documentation was removed and the
proper instructions about using a SAO DS9 region file are explained in the
'--polygon' option.