Library (tiff.h): corrected usage of uint32 with uint32_t
Until now, in two instances within the 'lib/tiff.c' library, we were using
the 'uint32' type (which is not the standard and not guaranteed to work on
all systems)! This happened because these functions started-off as a
copy/pasting from the example codes in the TIFF library. I noticed these
with compiler warnings in a recent build.

With this commit, they have been changed to the standard 'uint32_t'.