Book: edited gal_ds9_reg_read_polygon from DS9 Library, corrected NEWS
Until now, the function 'gal_ds9_reg_read_polygon' from the DS9 Library had
'*coordmode' as input to save the mode of the coordinates read from the DS9
region file. Also, the NEWS file still mentioned the '--polygonfile'
option!

With this commit, the change removing '*coordmode' as input and writing the
coordinate mode into the 'status' component of the output 'gal_data_t' is
now documented. Also, the NEWS file now includes this new feature under the
'--polygon' option.

Finally, the formatting of the value of this option in 'args.h' and also in
the book have been corrected. In 'args.h', we simply use 'FLT,FLT[:...]' to
avoid confusion, and say that a DS9 region file is also possible in the
description. In the book, there are two '--polygon' items: one with 'STR'
and one with 'FLT,FLT[:...]'.