Match: warning printed when redundant --ccol2 is given with --coord
Until now, when the '--coord' option was called with '--ccol2', Match would
crash with an ugly segmentation fault. The reason for this was that When
Match is called with '--coord', the table input is assumed to be the first
catalog, therefore there is no need for '--ccol2'. So I was free-ing its
value pre-maturely (while it is needed to print with something like '-P' or
in the FITS headers).

With this commit, instead of freeing the allocated space (and causing a lot
of complications), a warning will be printed for the user to tell them that
'--ccol2' is redundant.