astscript-ds9-region: now properly using --hdu
Until now, even though the '--hdu' option value was being read by this
script, its value was never used when reading the input FITS file!

With this commit, we now use the '--hdu' option any time that we read the
input from a file (not stdin).

I also noticed that the short version of the '--help' option in the output
of '--help' itself was mistakenly written as '-h' (similar to '--hdu')! It
has been corrected to the proper value ('-?'). Then I had a look at the
other scripts and noticed this problem existed in all of them! So they are
all corrected now.