Option parsing: warning printed when --config value doesn't exist
Until now, when the user specified a configuration file (with '--config')
and the file doesn't exist, Gnuastro programs wouldn't crash or print any
warning, but they would ignore it and continue quietly. This could lead to
bad confusions when there was a problem in readin the file (e.g., a typo in
the file name, change of location, no permission, or etc).

This is the expected behavior for default configuration files (we don't
want to bother the user about non-existent default directory or user
configuration files). But it shouldn't be the case for user-given
configuration files.

With this commit a warning will now be printed when a file given to the
'--config' option can't be opened for any reason. But because configuration
files are generally optional, the program won't abort. Hopefully the
warning will be clear enough.

This bug was reported by Sepideh Eskandarlou.