Auto-completion: clean re-implementation for Table
Until now, Pedram had made great progress with creating a beautiful core
infrastructure and developer documentation for the autocompletion. It was
so well written that I was easily able to experiment and go through the
model and understand its inner workings. In order to generalize it for any
generic Gnuastro program and make it scalable to many options, I noticed
that it is easier to do a clean re-implementation of the code.

With this commit, the first full re-implementation of the completion code
has been written for Table following the modularity and scalability
principles (to easily extend to other programs). It is now clearly
separated into parts that should be general for all programs and parts that
are custom built for each individual program.