Fits: four new options to find image/table HDUs
Until now, when a user wanted to see if at least on table/image HDU exists
in a FITS file, they would have to look at the full list of HDUs. For a
human eye that is fine, but for a script, this was annoying and slow. In
particular, as part of current effort to add shell auto-complete features
in Gnuastro. The same goes for scenarios when you want the list of image or
table HDUs.

With this commit, four new options have been added to the Fits program to
greatly simplify these jobs: '--hastablehdu', '--hasimagehdu',
'listtablehdu', '--listimagehdu'. The first two will only parse enough HDUs
to see if atleast one exists and the other two will print each
corresponding HDU name or number in a separate line.