Auto complete: separated generic functions from Table's recipe
Until now we were in development phase and thus only had the completion
script for the Table program. But now that things are tested enough and
ready, we can implement it in other programs too. So it was necessary to
separate the Gnuastro-generic from the program-specific functions.

With this commit, this has been done: the generic file is placed in
'bin/gnuastro-complete.bash' and the Table file has been placed in
'bin/table/asttable-complete.bash'. To help in understanding the process,
an extended comment is added on top of 'gnuastro-complete.bash' and to help
in debugging a (by default commented) line is added in
'asttable-complete.bash' to 'source' the generic functions.