TAB completion: now working for Arithmetic, better infrastructure
Until now, the Bash TAB completion was primarily designed around the table
program and still needed some polishing to be generalized. There was also
no documentation for the TAB completion in the book.

With this commit, during the inclusion of the TAB completion for
Arithmetic, things were generalized and are now also used in the Arithmetic
program. Also, since the Arithmetic program takes operators as command-line
arguments, it was necessary to also print the operators on the
command-line. Finally, the book was also updated with a full section
devoted to this.

I also noticed that the '--help' output of the '--envseed' option in
Arithmetic was based on the Table program! This has also been corrected.