TAB completion: support added for CosmicCalculator and Crop
With this commit, support for TAB completion in CosmicCalculator and Crop
have been added. Also, to help in the readability of the "MAYBE_" variables
related to completion in the top 'Makefile.am', their name is now in the
format of 'MAYBE_COMPLETE_*' (until now it was 'MAYBE_*_COMPLETE').

During the process, I noticed a segmentation fault in Crop (caused when a
single image was given, but no '--width' specified) and that 'astcrop -P'
will complain about no crop-mode being specified. Both these issues are
also fixed in this commit.