TAB completion: updated documentation and polished the code
Until now, major changes had occurred in the coding of the TAB completion
in Gnuastro, but the changes hadn't been reflected in the respective
section of "Developing" chapter of the Gnuastro book.

With this commit, those two sections were fully re-written to reflect the
recent changes. I should also thank Alfred and Richard (from the
'gnu-prog-discuss' mailing list) for their great suggestions:

 - Alfred Szmidt kindly suggested (in a previous commit) that the
   completion script be installed in 'pkgdatadir' and that the final
   printed message go into 'install-data-hook'.

 - Richard Wilbur reported a bug in the code example to active Bash TAB
   completion in the "Quick start" section (it should have used '>>', with
   the way I had written it, the whole '~/.bashrc' would be over-written.