Book: new section on known issues with Crop
Until now, there was only a "Known issues" section for the whole Gnuastro
installation step. But in other cases, it may also happen that programs
crash due to problems that are beyond Gnuastro itself (for example in
dependency libraries) and thus we can't do anything about it! This
situation happened recenty with a crash in the Crop program that was due to
CFITSIO.

With this commit, a new "Crop known issues" section has been added under
the "Invoking Crop" section to describe that problem and propose
solutions. If other similar problems occur, from now on, we'll have a place
to let the users know.

This issue was reported by Sepideh Eskandarloo.