Table: an empty input table will not cause a crash with --range
Until now, when Table was called with an empty input and the '--range'
option, it would crash with a Segmentation fault. This happened because we
were implicitly always assuming that the input table is full when '--range'
has been called (since it is the first row-limiter!).

With this commit, a check has been added at the start of the
'table_select_by_value' function so when the input is empty, it just
ignores the function over all.

This bug was reported by Sepideh Eskandarloo.