All programs: empty string arguments are now ignored
Until now, when an empty shell variable was given to any of the programs,
they would crash by saying a file called '' (empty string) doesn't
exist. This can happen a lot in a shell script, but this error isn't clear
and even confusing! A more natural error is that "no input file
specified". I noticed this most recently while working on the final changes
in the radial profile script.

With this commit, before writing the argument inside any of the programs,
they will first check to see if its an empty string or not. If it is indeed
empty, they will ignore the argument, thus letting the program abort with a
"no input" error.